requirements.txt
aiosqlite>=0.21.0, <1.0.0
httpx>=0.28.1, <0.29.0
aiohttp>=3.12.4, <4.0.0
python_dotenv>=1.1.0, <2.0.0
azure-identity>=1.23.0, <2.0.0
azure-ai-projects==1.0.0b11
azure-ai-agents==1.0.0
pandas>=2.2.3, <3.0.0
pydantic==2.11.5
pillow>=11.2.1, <12.0.0